refactor: replace AuthContext with UserContext across components and tests
