--- client/src/__testUtils__/fetchMocks.js
@@ -1,3 +1,0 @@
-export const asSlowResponse = (response) => {
-  return () => new Promise((resolve) => setTimeout(() => resolve(response)));
-};
