--- client/src/components/Nav.jsx
@@ -1,18 +1,0 @@
-import { Link } from "react-router-dom";
-
-import TEST_ID from "./Nav.testid";
-
-const Nav = () => {
-  return (
-    <ul>
-      <Link to="/" data-testid={TEST_ID.linkToHome}>
-        <li>Home</li>
-      </Link>
-      <Link to="/user" data-testid={TEST_ID.linkToUsers}>
-        <li>Users</li>
-      </Link>
-    </ul>
-  );
-};
-
-export default Nav;
