--- client/src/components/Nav.testid.js
@@ -1,8 +1,0 @@
-import createTestIdFilePath from "../util/createTestIdFilePath";
-
-const TEST_ID = {
-  linkToHome: `${createTestIdFilePath("components", "Nav")}-linkToHome`,
-  linkToUsers: `${createTestIdFilePath("components", "Nav")}-linkToUser`,
-};
-
-export default TEST_ID;
