+++ client/src/components/SkillsSettings.jsx
@@ -3,12 +3,12 @@ import { validateSkillInput } from "../util/skillValidation";
 import AlertMessage from "./AlertMessage";
 import { regexEndNormalizeSkill } from "../util/regexEndNormalizeSkill";
 import { cleanUpText } from "../util/cleanUpText";
-import { UseAuth } from "../context/AuthContext";
+import { UseUser } from "../context/UserContext";
 
 export default function SkillsSettings() {
   const skillInputRef = useRef(null);
   const [alert, setAlert] = useState({ type: "", message: "" });
-  const { user, setUser } = UseAuth();
+  const { user, setUser } = UseUser();
   const { skills } = user;
 
   function handleClearAlert() {
