+++ client/src/main.jsx
@@ -4,22 +4,16 @@ import App from "./App.jsx";
 import { StrictMode } from "react";
 import { BrowserRouter } from "react-router-dom";
 import { JobsProvider } from "./context/JobsContext.jsx";
-// import { FavoritesProvider } from "./context/FavoritesContext.jsx";
-// import { SettingsProvider } from "./context/SettingsContext.jsx";
-import { AuthProvider } from "./context/AuthContext.jsx";
+import { UserContextProvider } from "./context/UserContext.jsx";
 
 createRoot(document.getElementById("root")).render(
   <StrictMode>
     <BrowserRouter>
-      {/* <SettingsProvider> */}
-      <AuthProvider>
+      <UserContextProvider>
         <JobsProvider>
-          {/* <FavoritesProvider> */}
           <App />
-          {/* </FavoritesProvider> */}
         </JobsProvider>
-      </AuthProvider>
-      {/* </SettingsProvider> */}
+      </UserContextProvider>
     </BrowserRouter>
   </StrictMode>,
 );
