+++ client/src/pages/LoginForm/LoginForm.jsx
@@ -1,10 +1,10 @@
 import { useState } from "react";
 import { Loader, X, LogIn } from "lucide-react";
-import { UseAuth } from "../../context/AuthContext";
+import { UseUser } from "../../context/UserContext";
 
 const LoginForm = ({ login, switchToSignup }) => {
   const [loginData, setLoginData] = useState({ email: "", password: "" });
-  const { loading, error, clearError } = UseAuth();
+  const { loading, error, clearError } = UseUser();
 
   const handleChange = (e) => {
     const { name, value } = e.target;
