+++ client/src/pages/User/Profile.jsx
@@ -5,7 +5,7 @@ import AlertMessage from "../../components/AlertMessage";
 import { cleanUpText } from "../../util/cleanUpText";
 import { validateAddressTextInputs } from "../../util/addressTextsValidation";
 import { validateHouseNoInput } from "../../util/addressHouseNoValidation";
-import { UseAuth } from "../../context/AuthContext";
+import { UseUser } from "../../context/UserContext";
 
 export default function Profile() {
   const [alert, setAlert] = useState({ type: "", message: "" });
@@ -17,7 +17,7 @@ export default function Profile() {
   const houseInputRef = useRef(null);
   const cityInputRef = useRef(null);
   const countryInputRef = useRef(null);
-  const { user, setUser } = UseAuth();
+  const { user, setUser } = UseUser();
 
   function handleClearAlert() {
     if (!alert.message) return;
