--- client/src/util/__tests__/createTestIdFilePath.test.js
@@ -1,7 +1,0 @@
-import createTestIdFilePath from "../createTestIdFilePath";
-
-describe("createTestIdFilePath", () => {
-  it("Adds all of the args", () => {
-    expect(createTestIdFilePath("a", "b", "c")).toBe("a/b/c");
-  });
-});
